rename main package
