+++ src/main/scala/org/jsonschematools/merging/MMerge.scala
@@ -1,4 +1,4 @@
-package org.example.merging
+package org.jsonschematools.merging
 
 import org.json4s._
 import org.json4s.JNothing
