+++ src/main/scala/org/jsonschematools/schema/types/JsonSchemaTypeNotFound.scala
@@ -1,3 +1,3 @@
-package org.example.schema.types
+package org.jsonschematools.schema.types
 
 case class JsonSchemaTypeNotFound(msg: String) extends Exception(msg)
